Dodate metode poslovne logike

+++ BusinessLayer/DataLayer/Models/Student.cs
@@ -11,7 +11,7 @@ namespace DataLayer.Models
         public int Id { get; set; }
         public string Name { get; set; }
         public string IndexNumber { get; set; }
-        public float AverageMark { get; set; }
+        public decimal AverageMark { get; set; }
 
     }
 }
